Quiet warnings. Add all the Boost headers to the Xcode project. More migration to NSInteger.

+++ macintosh/sceneDocument/colorPicker.mm
@@ -187,7 +187,7 @@
 //---------------------------------------------------------------------
 -(IBAction) colorPickerTraget:(id)sender
 {
-	int theTag;
+	NSInteger theTag;
 	if ( sender==self)
 		theTag=cFilterOn;
 	else
