+++ macintosh/sceneDocument/sceneDocument.mm
@@ -658,7 +658,7 @@
 	{
 		if ( [[mSceneTextView textStorage]length]/(1024*1024) > 25l)
 		{
-			int res=NSRunAlertPanel( NSLocalizedStringFromTable(@"DocumentIsHuge", @"applicationLocalized", @"This is a large file\n"),
+			NSInteger res=NSRunAlertPanel( NSLocalizedStringFromTable(@"DocumentIsHuge", @"applicationLocalized", @"This is a large file\n"),
 														NSLocalizedStringFromTable(@"DocumentIsHuge2", @"applicationLocalized", @"Turning on syntax coloring will lock up the computer for a long time.\nDo you want to proceed anyway?"),
 														NSLocalizedStringFromTable(@"ColorOff", @"applicationLocalized", @"Color off"),
 														NSLocalizedStringFromTable(@"ColorOn", @"applicationLocalized", @"Color on"),
@@ -853,7 +853,7 @@
 				NSString *hugeString=[NSString stringWithString:[absoluteURL lastPathComponent]];
 				hugeString=[hugeString stringByAppendingString:messageString];
 				hugeString =[hugeString stringByAppendingString:@"25Mb."];
-				int res=NSRunAlertPanel( hugeString,
+				NSInteger res=NSRunAlertPanel( hugeString,
 																NSLocalizedStringFromTable(@"StillLoad", @"applicationLocalized", @"Load with color off, color on or skip loading?"),
 																	NSLocalizedStringFromTable(@"DontLoad", @"applicationLocalized", @"Skip"),
 																NSLocalizedStringFromTable(@"ColorOn", @"applicationLocalized", @"Color on"),
@@ -1410,7 +1410,7 @@ NS_ENDHANDLER
 				line ++;
 		}
 	}
-	[gotoEdit setIntValue:line];
+	[gotoEdit setIntegerValue:line];
 	[mSceneTextView setTypingAttributes:[[appPreferencesController sharedInstance] style: cBlackStyle ]];
 
 }
